+++ Sparkle/src/widget/components/spk_nine_sliced.cpp
@@ -64,7 +64,7 @@ namespace spk::WidgetComponent
         {
             for (size_t y = 0; y < 3; y++)
             {
-                unsigned int baseIndexes = _bufferShaderInput.size();
+                unsigned int baseIndexes = static_cast<unsigned int>(_bufferShaderInput.size());
                 for (size_t i = 0; i < 2; i++)
                 {
                     for (size_t j = 0; j < 2; j++)
